+++ sonic_number/README.md
@@ -1,0 +1,3 @@
+# sonic_number
+
+A fast number parsing library based on SIMD.
