+++ sonic_number/src/lemire.rs
