+++ sonic_simd/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+edition = "2021"
+name    = "sonic_simd"
+version = "0.1.0"
+
+[dependencies]
+cfg-if = "1.0"
