+++ sonic_simd/README.md
@@ -1,0 +1,8 @@
+
+# sonic_simd
+
+A portable SIMD library that provides low-level APIs for x86, ARM. Other platforms will use the fallback scalar implementation.
+
+TODO:
+
+1. support RISC-V.
