+++ sonic_simd/src/neon.rs
@@ -1,7 +1,4 @@
-use std::{
-    arch::aarch64::*,
-    ops::{BitAnd, BitOr, BitOrAssign},
-};
+use std::arch::aarch64::*;
 
 use super::{bits::NeonBits, Mask, Simd};
 
@@ -151,12 +148,7 @@ impl std::ops::BitOrAssign<Mask128> for Mask128 {
 }
 
 #[inline(always)]
-pub(crate) unsafe fn to_bitmask64(
-    v0: uint8x16_t,
-    v1: uint8x16_t,
-    v2: uint8x16_t,
-    v3: uint8x16_t,
-) -> u64 {
+pub unsafe fn to_bitmask64(v0: uint8x16_t, v1: uint8x16_t, v2: uint8x16_t, v3: uint8x16_t) -> u64 {
     let bit_mask = std::mem::transmute::<[u8; 16], uint8x16_t>(BIT_MASK_TAB);
 
     let t0 = vandq_u8(v0, bit_mask);
