+++ sonic_simd/src/sse2.rs
