+++ src/util/mod.rs
@@ -1,15 +1,10 @@
 pub(crate) mod arch;
 
-pub(crate) mod num;
 pub(crate) mod private;
 pub(crate) mod reborrow;
 pub(crate) mod string;
 pub(crate) mod unicode;
 pub(crate) mod utf8;
-
-#[allow(non_camel_case_types)]
-#[allow(unused_imports)]
-pub(crate) mod simd;
 
 #[cfg(test)]
 pub(crate) mod mock;
